requirements.txt
# 核心依赖
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
numpy>=1.24.0
pillow>=10.0.0
requests>=2.31.0

# 向量数据库（用于上下文存储）
chromadb>=0.4.0
sentence-transformers>=2.2.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0

# 可选：用于评估和测试
pytest>=7.4.0
